Blog_Pribadi/requirements.txt
fastapi[standard]>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
pymysql>=1.1.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6